+++ layers/GatoGraphQLForWP/plugins/gatographql/src/ModuleConfiguration.php
@@ -282,4 +282,17 @@ class ModuleConfiguration extends AbstractModuleConfiguration
             $callback,
         );
     }
+
+    public function enableReverseLogOrder(): bool
+    {
+        $envVariable = Environment::ENABLE_REVERSE_LOG_ORDER;
+        $defaultValue = false;
+        $callback = EnvironmentValueHelpers::toBool(...);
+
+        return $this->retrieveConfigurationValueOrUseDefault(
+            $envVariable,
+            $defaultValue,
+            $callback,
+        );
+    }
 }